refactor: state, actions, feedbacks
