+++ src/actions.ts
@@ -11,8 +11,8 @@ import { EmberClient, Model as EmberModel } from 'emberplus-connection'
 import type PQueue from 'p-queue'
 import type { EmberPlusConfig } from './config'
 import type { EmberPlusInstance } from './index'
-import { doMatrixAction, doTake, doClear, setSelectedSource, setSelectedTarget } from './matrix'
-import { learnSetValueActionOptions, setValue, subscribeParameterAction } from './parameter'
+import { doMatrixAction, doTake, doClear, setSelectedSource, setSelectedTarget } from './actions/matrix'
+import { learnSetValueActionOptions, setValue, subscribeParameterAction } from './actions/parameter'
 import { EmberPlusState } from './state'
 import { filterPathChoices } from './util'
